Make the --verbose flag universal
Previously, the --verbose flag was only present in debug builds. Now
it's present whether you define DEBUG or not.  What changes is the
amount of information printed to stderr: DEBUG sets the --verbose log
level to 0 (debug), while DEBUG unset sets it to 1 (info).  This makes
driving the binary slightly simpler as you don't have to detect whether
it's a debug build by scanning for "--verbose" in the help output.